Organized home and added temp endpoint for onboarding
