+++ pubspec.yaml
@@ -41,6 +41,7 @@ dependencies:
   firebase_auth: ^4.1.3
   provider: ^6.0.4
   http: ^0.13.5
+  firebase_database: ^10.0.6
 
 dev_dependencies:
   flutter_test:
